% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgGEMset-class.R
\docType{class}
\name{ctgGEMset}
\alias{ctgGEMset}
\alias{.ctgGEMset}
\alias{ctgGEMset-class}
\title{The ctgGEMset class}
\usage{
ctgGEMset(exprsData, phenoData = NULL, featureData = NULL)
}
\arguments{
\item{exprsData}{expression data matrix for an experiment}

\item{phenoData}{a data frame containing attributes of individual samples}

\item{featureData}{a data frame containing attributes of features (genes)}
}
\value{
a new ctgGEMset object
}
\description{
The main class used by the \pkg{cellTreeGenerator} package to hold single
cell gene expression data and generated tree results.  ctgGEMset extends the
\link[SummarizedExperiment:SummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} class.
}
\details{
This class is initialized from a matrix of gene expression values and
associated metadata.  Methods that operate on ctgGEMset objects
comprise the ctgGEM workflow.
}
\section{Fields}{

\describe{
\item{\code{cellTreeInfo}}{a character vector of the name of the column of
\code{phenoData()} to use for grouping data for
\code{generate_tree(treeType = "cellTree")} in the
\pkg{cellTreeGenerator} workflow}

\item{\code{monocleInfo}}{A character vector of parameters used by
\code{generate_tree(treeType = "monocle")} in the
\pkg{cellTreeGenerator} workflow}

\item{\code{TSCANinfo}}{A character vector of the row name of a single gene in
\code{exprsData()} to use for a single gene vs. pseudotime plot for
\code{generate_tree(treeType = "TSCAN")} in the
\pkg{cellTreeGenerator} workflow}

\item{\code{sincellInfo}}{A list containing named parameters used by
\code{generate_tree(treeType = "sincell")} in the
\pkg{cellTreeGenerator} workflow}

\item{\code{treeList}}{A list containing the simplified igraph representation of
the trees generated by the \pkg{ctgGEM} workflow}

\item{\code{originalTrees}}{A list containing the trees generated by the
\pkg{ctgGEM} workflow in their original formats for re-plotting}
}}

\examples{
# load HSMMSingleCell package
library(HSMMSingleCell)

# load the data
data(HSMM_expr_matrix)
data(HSMM_sample_sheet)
data(HSMM_gene_annotation)

# construct a ctgGEMset
dataSet <- ctgGEMset(exprsData = HSMM_expr_matrix,
                        phenoData = HSMM_sample_sheet,
                        featureData = HSMM_gene_annotation)
# load HSMMSingleCell package
library(HSMMSingleCell)

# load the data
data(HSMM_expr_matrix)
data(HSMM_sample_sheet)
data(HSMM_gene_annotation)


# construct a ctgGEMset
dataSet <- ctgGEMset(exprsData = HSMM_expr_matrix,
                        phenoData = HSMM_sample_sheet,
                        featureData = HSMM_gene_annotation)
}
