% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_tree}
\alias{generate_tree}
\title{Function to Generate Cell Trees}
\usage{
generate_tree(dataSet, treeType)
}
\arguments{
\item{dataSet}{the ctgGEMset object for creating the cell tree}

\item{treeType}{the type of tree generated}
}
\value{
An updated ctgGEMset object.  The generated tree is placed in
\code{@treeList[treeType]} slot, and can be accessed via
\code{treeList(dataSet)$treeType}.  The function also creates a
directory named "treeType-Output" and writes the plot(s) of the
generated tree(s) and its SIF file to that directory.
}
\description{
This function builds a cell hierarchy tree of a chosen supported type with
a given data set, contained in a ctgGEMset object.  Different tree
types require data from corresponding slots of the ctgGEMset object.
See vignette for examples, usage details, and instructions on building a
ctgGEMset object.
}
\examples{
# load HSMMSingleCell package
library(HSMMSingleCell)

# load the data
data(HSMM_expr_matrix)
data(HSMM_sample_sheet)
data(HSMM_gene_annotation)

# construct a ctgGEMset
dataSet <- newctgGEMset(exprsData = HSMM_expr_matrix,
                        phenoData = HSMM_sample_sheet,
                        featureData = HSMM_gene_annotation)

TSCANinfo(dataSet) <- "ENSG00000000003.10"

# run generate_tree()
dataSet <- generate_tree(dataSet = dataSet, treeType = "TSCAN")
}
\keyword{cell}
\keyword{tree}
