% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{make_method_matrix}
\alias{make_method_matrix}
\title{A Function That Creates A Predefined Matrix Of Usable Packages}
\usage{
make_method_matrix()
}
\value{
a matrix containing all possible packages and their corresponding
function
}
\description{
This function constructs the predefined matrix that contains all possible
cell tree building packages and their corresponding functions
}
\keyword{internal}
