% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCellTree.R
\name{makeCellTree}
\alias{makeCellTree}
\title{A Cell Tree Generating Function using cellTree}
\usage{
makeCellTree(dataSet)
}
\arguments{
\item{dataSet}{a ctgGEMset object}
}
\value{
an updated ctgGEMset object
}
\description{
This function, called by \code{\link{generate_tree}}, creates a cell tree
using the data and the \pkg{cellTree} package. This function utilizes code
from the cellTree vignette R script.
}
\keyword{internal}
