% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDestiny.R
\name{makeDestiny}
\alias{makeDestiny}
\title{A Cell Tree Generating Function using destiny}
\usage{
makeDestiny(dataSet)
}
\arguments{
\item{dataSet}{a ctgGEMset object}
}
\value{
an updated ctgGEMset object
}
\description{
This function, called by \code{\link{generate_tree}}, creates visualizations
using the data using the \pkg{destiny} package. This function utilizes code
from the destiny and dpt vignette R scripts.
}
\keyword{internal}
