% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTSCAN.R
\name{makeTSCAN}
\alias{makeTSCAN}
\title{A Cell Tree Generating Function using TSCAN}
\usage{
makeTSCAN(dataSet, outputDir = tempdir())
}
\arguments{
\item{dataSet}{a ctgGEMset object}

\item{outputDir}{the directory where output should be saved, defaults to
the temporary location returned by \code{tempdir()}}
}
\value{
an updated ctgGEMset object
}
\description{
This function, called by \code{\link{generate_tree}}, creates a cell tree
using the data and the \pkg{TSCAN} package. This function utilizes code
from the TSCAN vignette R script.
}
\keyword{internal}
