% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctgGEMset-class.R
\name{newctgGEMset}
\alias{newctgGEMset}
\title{Creates a new ctgGEMset object.}
\usage{
newctgGEMset(exprsData, phenoData = NULL, featureData = NULL)
}
\arguments{
\item{exprsData}{expression data matrix for an experiment}

\item{phenoData}{a data frame containing attributes of individual samples}

\item{featureData}{a data frame containing attributes of features (genes)}
}
\value{
a new ctgGEMset object
}
\description{
Creates a new ctgGEMset object.
}
\examples{
# load HSMMSingleCell package
library(HSMMSingleCell)

# load the data
data(HSMM_expr_matrix)
data(HSMM_sample_sheet)
data(HSMM_gene_annotation)


# construct a ctgGEMset
dataSet <- newctgGEMset(exprsData = HSMM_expr_matrix,
                        phenoData = HSMM_sample_sheet,
                        featureData = HSMM_gene_annotation)
}
