% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOriginalTree.R
\name{plotOriginalTree}
\alias{plotOriginalTree}
\title{Display Original ctgGEM Plots}
\usage{
plotOriginalTree(dataSet, treeType)
}
\arguments{
\item{dataSet}{a ctgGEMset object}

\item{treeType}{the type of tree to display.  Must be one of
\code{names(originalTrees(dataSet))}}
}
\value{
For \code{treeType = "destinyDM"} and
\code{treeType = "destinyDPT"}, returns \code{NULL}.
For all other tree types, a ggplot2::ggplot object.
}
\description{
Displays the original plots created by the ctgGEM package and
stored in the \link{originalTrees} slot of a ctgGEMset object.
}
\note{
In order to reproduce original oncoNEM trees or destiny plots, the
respective package(s) must be installed.
}
\examples{
# load HSMMSingleCell package
library(HSMMSingleCell)

# load the data for TSCAN and monocle:
data(HSMM_expr_matrix)
data(HSMM_sample_sheet)
data(HSMM_gene_annotation)

# construct a ctgGEMset
dataSet <- ctgGEMset(exprsData = HSMM_expr_matrix,
                        phenoData = HSMM_sample_sheet,
                        featureData = HSMM_gene_annotation)

TSCANinfo(dataSet) <- "ENSG00000000003.10"

# run generate_tree()
dataSet <- generate_tree(dataSet = dataSet, treeType = "TSCAN")

# view names of original trees
names(originalTrees(dataSet))

# plot original trees
plotOriginalTree(dataSet, "TSCANclustering")
plotOriginalTree(dataSet, "TSCANsingleGene")
}
