% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree2igraph.R
\name{tree2igraph}
\alias{tree2igraph}
\title{Convert Cell Relationships in a SIF object to igraph Format}
\usage{
tree2igraph(relationships)
}
\arguments{
\item{relationships}{character vector of a cell tree in SIF format}
}
\value{
a cell tree in igraph format
}
\description{
This function converts cell relationships in a cell tree into igraph format
}
\keyword{internal}
