% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMonocle.R
\name{makeMonocle}
\alias{makeMonocle}
\title{A Cell Tree Generating Function using monocle}
\usage{
makeMonocle(dataSet)
}
\arguments{
\item{dataSet}{a ctgGEMset object}
}
\value{
an updated ctgGEMset object
}
\description{
This function, called by \code{\link{generate_tree}}, creates a cell tree
using the \pkg{monocle} package. This function utilizes code
from the monocle vignette R script.
}
\keyword{internal}
